Added JSDOM so we can search for selectors.
Added JSDOM so we can search for CSS selectors on the page.

Fixed outputs so that if a flag is used, that will be added to the output. This is different from previous behavior as we would only do one of the flags. Now we do any variation of them if they're used.